refactor: Simplify EventListDialog by delegating to specialized components
- Refactor EventListDialog.kt to act as a facade, delegating event handling and dialog management to specialized components for improved maintainability and organization.
- Remove legacy UI logic and streamline the structure for better clarity and consistency.
